+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2020-08-20
+### Added
+- Added support for train case.
+- Add new JSON key decoding strategies.
+- Added `convert` function for converting from one lettercase to another.
+
 ## [1.3.1] - 2020-08-19
 ### Changed
 - Moved KeyDecodingStrategyAdditions for SPM.